requirements.txt
openai>=1.0.0
pydantic>=2.5
fastapi>=0.104
uvicorn[standard]>=0.24
streamlit>=1.28
pyyaml>=6.0
python-dotenv>=1.0
groq>=0.9.0